Utility scripts for base provisioning
You could do things like installing the VirtualBox guest additions in
Puppet, but these scripts exist for provisioning a box that doesn't
have Puppet on it. There is definitely a legit use case for these
'nocm' boxes. Most boxes you'll be given in production won't have
Puppet on them, so they help to try and anticipate everything that you
can.